vector-store:test: extract TLS test utilities
Extract TLS test utilities for reuse across mTLS test suites
